refactor: 이벤트 패키지 구조를 모듈 기준으로 정리
