--- src/main/java/com/flab/woowahaneats/domain/order/event/OrderAcceptedEvent.java
@@ -1,6 +1,0 @@
-package com.flab.woowahaneats.domain.order.event;
-
-import java.util.UUID;
-
-public record OrderAcceptedEvent(UUID userOrderId) {
-}
