+++ src/main/java/com/flab/woowahaneats/domain/order/owner/event/OwnerOrderAcceptedEvent.java
@@ -1,0 +1,6 @@
+package com.flab.woowahaneats.domain.order.owner.event;
+
+import java.util.UUID;
+
+public record OwnerOrderAcceptedEvent(UUID userOrderId) {
+}
