+++ src/main/java/com/flab/woowahaneats/domain/order/user/application/UserOrderService.java
@@ -10,7 +10,6 @@ import com.flab.woowahaneats.domain.member.domain.User;
 import com.flab.woowahaneats.domain.menu.application.exception.MenuNotFoundException;
 import com.flab.woowahaneats.domain.menu.domain.Menu;
 import com.flab.woowahaneats.domain.menu.repository.MenuRepository;
-import com.flab.woowahaneats.domain.order.event.UserOrderCancelledEvent;
 import com.flab.woowahaneats.domain.order.user.controller.dto.CreateOrderResponse;
 import com.flab.woowahaneats.domain.payment.application.PaymentService;
 import com.flab.woowahaneats.domain.payment.domain.Payment;
@@ -24,6 +23,7 @@ import com.flab.woowahaneats.domain.order.user.controller.dto.CreateOrderRequest
 import com.flab.woowahaneats.domain.order.user.controller.dto.OrderResponse;
 import com.flab.woowahaneats.domain.order.user.domain.UserOrder;
 import com.flab.woowahaneats.domain.order.common.OrderMenu;
+import com.flab.woowahaneats.domain.order.user.event.UserOrderCancelledEvent;
 import com.flab.woowahaneats.domain.order.user.repository.UserOrderRepository;
 import com.flab.woowahaneats.domain.restaurant.domain.RestaurantOperationInfo;
 import com.flab.woowahaneats.domain.restaurant.repository.RestaurantOperationInfoRepository;
