+++ src/main/java/com/flab/woowahaneats/domain/payment/application/PaymentService.java
@@ -2,11 +2,11 @@ package com.flab.woowahaneats.domain.payment.application;
 
 import com.flab.woowahaneats.domain.auth.AuthContextHolder;
 import com.flab.woowahaneats.domain.member.domain.User;
-import com.flab.woowahaneats.domain.order.event.PaymentCompletedEvent;
 import com.flab.woowahaneats.domain.order.user.domain.UserOrder;
 import com.flab.woowahaneats.domain.order.user.repository.UserOrderRepository;
 import com.flab.woowahaneats.domain.payment.domain.Payment;
 import com.flab.woowahaneats.domain.payment.domain.PaymentProvider;
+import com.flab.woowahaneats.domain.payment.event.PaymentCompletedEvent;
 import com.flab.woowahaneats.domain.payment.exception.PaymentApprovalFailedException;
 import com.flab.woowahaneats.domain.payment.exception.PaymentNotBelongToUserException;
 import com.flab.woowahaneats.domain.payment.exception.PaymentNotFoundException;
